- Added a gem file to run tests. - Added the utils app - Got some tests working
